Add files via upload
This presentation is a comprehensive overview of Digital Signatures, a secure way to authenticate and verify the integrity of digital messages and documents. It covers the fundamental concepts of traditional signatures, electronic signatures, and dives deep into digital signatures. The presentation explains how digital signatures work, the benefits of using them, and the cryptographic techniques involved. Attendees will understand the role of certificate authorities (CAs) and the process of obtaining and verifying digital signatures. This presentation is part of the term work for the Software Security course taught by Dr. Basheer Abdel Fatah Youssef at the Faculty of Computers and Artificial Intelligence, Cairo University.